@@ -11,19 +11,24 @@
   },
   "devDependencies": {},
   "dependencies": {
-    "@strapi/strapi": "4.11.3",
-    "@strapi/plugin-users-permissions": "4.11.3",
-    "@strapi/plugin-i18n": "4.11.3",
-    "better-sqlite3": "8.0.1"
+    "@strapi/strapi": "4.20.4",
+    "@strapi/plugin-users-permissions": "4.20.4",
+    "@strapi/plugin-i18n": "4.20.4",
+    "@strapi/plugin-cloud": "4.20.4",
+    "better-sqlite3": "8.6.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "5.3.4",
+    "styled-components": "5.3.3"
   },
   "author": {
     "name": "A Strapi developer"
   },
   "strapi": {
-    "uuid": "1def158a-38a8-4d50-a23a-40430efc8a84"
+    "uuid": "263154f4-8080-4396-a572-06017626b7c1"
   },
   "engines": {
-    "node": ">=14.19.1 <=18.x.x",
+    "node": ">=18.0.0 <=20.x.x",
     "npm": ">=6.0.0"
   },
   "license": "MIT"
