@@ -1,6 +1,6 @@
-const path = require('path');
+import path from 'path';
 
-module.exports = ({ env }) => {
+export default ({ env }) => {
   const client = env('DATABASE_CLIENT', 'sqlite');
 
   const connections = {
@@ -74,6 +74,7 @@ module.exports = ({ env }) => {
       connection: {
         filename: path.join(
           __dirname,
+          '..',
           '..',
           env('DATABASE_FILENAME', '.tmp/data.db')
         ),
