@@ -1,9 +1,9 @@
-module.exports = [
+export default [
+  'strapi::logger',
   'strapi::errors',
   'strapi::security',
   'strapi::cors',
   'strapi::poweredBy',
-  'strapi::logger',
   'strapi::query',
   'strapi::body',
   'strapi::session',
